refactor(checker): split element-access const-enum/negative-index guard out of access.rs to hold 2000 LOC
access.rs had accumulated to 2026 physical lines on main (bypass-merge debt),
tripping the hard 2000-line checker-source arch cap for every PR's arch-guard.
Extract the TS2476 const-enum-index and negative-tuple-index guard blocks out of
the get_type_of_element_access_with_request monolith into access_helpers.rs
(pub(crate)). The receiver-is-const-enum check is shared by both the guard and
the later TS2339 const-enum missing-member path via a small
element_access_receiver_is_const_enum bool helper (that is the only thing both
sites used). No behavior change.

access.rs 2026->1963; access_helpers.rs 1372->1449. No other checker src file
exceeds 2000. Does not touch the #8225 quarantine cap.
